Extract local copy planning types into dedicated module
